ECL.ClassicChecker

Documentation

type Case = [Binder]

Type Synonyms for pattern * representation *

type Cases = [Case]

type Result = Cases

type DefPm = Cases

unmatched :: Case -> Cases

unmatched: returns the list * of `_` for the first uncovered set

unmatched' :: Case -> Case

ccheck :: DefPm -> Result

check: given a pattern matching definition * we check whether or not it is exhaustive * giving the resultant non-covered cases

uncover :: Binder -> Binder -> Case

uncover: single binder uncovering * this returns the difference between * two single binders

uncovers :: Case -> Case -> Cases

uncovers: given an uncovered case and a clause * of a pattern match definition, * returns the uncovered cases

uncovering :: Cases -> Case -> Cases

uncovering: returns the cases * that are not covered by a single clause

check_uncovers :: Cases -> Cases -> Cases

check_uncovers: Given an uncovered set and a full pm definition * this calculates the cases that are not covered by * the pattern matching definition

is_exhaustive :: DefPm -> Bool

is_exhaustive: returns whether or not * the given definition is exhaustive